Consolidate local search into existing train.py and search_env.py
Address review feedback on PR #662: remove the duplicated search_env_local.py
and train_local.py and instead dispatch on a backend flag, following the
pattern in recipes/sdft.

- search_env.py: rename chroma_tool -> search_tool in SearchEnvGroupBuilder
  (accepts ChromaTool | LocalSearchTool — both expose the same @tool-decorated
  search method). Add LocalSearchR1DatasetBuilder alongside SearchR1DatasetBuilder.
- train.py: add backend: Literal['chroma', 'local'] = 'chroma' to CLIConfig
  and dispatch via small helper functions. Run name includes the backend.
- search_env_local.py, train_local.py: removed.
- README: point the lightweight section at 'train backend=local'.
